# openapi.yaml
openapi: 3.1.0
info:
  title: Food Bank API
  version: '1.0.0'
  description: |
    JSON API for running a food bank: referrals in, sessions scheduled, pick
    lists produced, stock adjusted.

    **Read `API.md` alongside this.** A schema describes shapes; it cannot
    describe the login/refresh cycle, what happens to a referral from an
    unrecognised address, or why a team lead receives fewer fields than an
    admin. Those are the parts most likely to be got wrong.

    ## Generating a client

    ```
    npx openapi-typescript openapi.yaml -o src/api/schema.d.ts
    ```
servers:
  - url: http://127.0.0.1:8787
    description: Local wrangler dev
  - url: https://foodbank-server-production.workers.dev
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/stock/levels:
    get:
      tags: [Stock]
      summary: Stock levels, ordered by shelf
      description: |
        The picking and stock-take screen. Ordered so a picker walks the aisle
        once: A1, A2, A10 — not alphabetically. Levels are derived by summing
        the ledger; there is no stored balance.
      parameters:
        - { in: query, name: includeInactive, schema: { type: string, enum: ['true'] } }
      responses:
        '200':
          description: Levels
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/StockLevel' }
  /api/v1/stock/search:
    get:
      tags: [Stock]
      summary: Autocomplete stock items
      description: Type `sug`, get `Sugar`. Prefix match first, then infix.
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 1, maxLength: 40 }
          description: >
            Capped at 40 characters — the database limits pattern length, and
            exceeding it is an error rather than an empty result.
      responses:
        '200':
          description: Matches, up to 20
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/StockItem' }
        '400': { description: Search term too long or empty }
  /api/v1/stock/items:
    get:
      tags: [Stock]
      summary: List stock items
      parameters:
        - { in: query, name: includeInactive, schema: { type: string, enum: ['true'] } }
      responses:
        '200':
          description: Items, by shelf
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/StockItem' }
    post:
      tags: [Stock]
      summary: Add a stock item
      description: Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              # Unit and low-stock controls are deliberately not part of stock maintenance.
              required: [name, shelfNumber]
              properties:
                name: { type: string, maxLength: 120 }
                shelfNumber: { type: string, maxLength: 20, examples: ['A1', '12b'] }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockItem' }
        '409': { description: An item with that name already exists }
  /api/v1/stock/items/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      tags: [Stock]
      summary: Amend a stock item
      description: 'Admin only. Set `isActive: false` to retire without losing history.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name: { type: string }
                shelfNumber: { type: string }
                isActive: { type: boolean }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockItem' }
  /api/v1/stock/take:
    post:
      tags: [Stock]
      summary: Save one page of a stock take
      description: |
        The weekly count is the **single source of truth**. For every item in
        the body, whatever the ledger held is discarded and replaced by one
        `opening_balance` at the counted figure — so the item's level afterwards
        *is* the number that was typed. There is no variance and nothing to
        reconcile.

        **Send only the items whose number the volunteer changed.** An item left
        out is not touched. That means an untouched item is either
        counted-and-correct or never counted, and the charity has accepted not
        being able to tell those apart — do not send unchanged rows to resolve
        it, because that would delete history they expect to keep.

        There is no stock take resource: no id, no open/commit lifecycle, no
        record that one started or finished. Each page save stands alone, which
        is why there is **no `409`** — there is no state to conflict with.

        **Repeating a save is safe.** The delete removes what the previous save
        wrote, so sending the same page twice leaves the same rows behind. It is
        last-write-wins rather than apply-once: a stale page saved late will
        overwrite a newer count, which is what `levels` in the response is for.

        A `countedQuantity` of `0` is legitimate and means the shelf is empty.
        It writes no ledger row at all — a zero delta is forbidden — and the
        level reads as zero because nothing is summed.

        Locking is by process: the charity arranges that nothing else changes
        during a count, and the server does not enforce it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [counts]
              properties:
                counts:
                  type: array
                  minItems: 1
                  maxItems: 200
                  description: >-
                    The changed items only. 200 is a cap on one request, not the
                    page size — the client chooses that, and the screen shows 40.
                  items:
                    type: object
                    required: [stockItemId, countedQuantity]
                    properties:
                      stockItemId: { type: string, format: uuid }
                      countedQuantity: { type: integer, minimum: 0, maximum: 100000 }
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema:
                type: object
                required: [applied, levels]
                properties:
                  applied: { type: integer }
                  levels:
                    type: array
                    description: The resulting level of each item in the request.
                    items:
                      type: object
                      required: [stockItemId, quantityOnHand]
                      properties:
                        stockItemId: { type: string, format: uuid }
                        quantityOnHand: { type: integer }
        '400':
          description: >-
            Empty `counts`, a quantity out of range, or the same `stockItemId`
            twice — two counts for one item is ambiguous and is refused rather
            than resolved.
        '404': { description: An unknown stock item. Nothing is written. }

  # ---------------------------------------------------------- parcel rules --
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        `Authorization: Bearer <accessToken>`. Expires after 15 minutes, or
        when the eight-hour sign-in ends if that comes first; see the refresh
        flow in API.md.
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    StockItem:
      type: object
      # Unit and low-stock controls are deliberately not part of stock maintenance.
      required: [id, name, shelfNumber, isActive]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        shelfNumber: { type: string }
        isActive: { type: boolean }
    StockLevel:
      allOf:
        - $ref: '#/components/schemas/StockItem'
        - type: object
          required: [quantityOnHand]
          properties:
            quantityOnHand:
              type: integer
              description: >-
                Derived by summing the ledger. **May be negative** — parcels go
                out between counts, and nothing stops an item going below what
                the last count said was there.
...